# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PSKit Research Agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/skills": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Skills",
        "description": "List Skills visible to the authenticated user.",
        "operationId": "list_skills_api_v1_skills_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogItem"
                  },
                  "type": "array",
                  "title": "Response List Skills Api V1 Skills Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/resources": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Resources",
        "description": "List resources visible to the authenticated user.",
        "operationId": "list_resources_api_v1_resources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogItem"
                  },
                  "type": "array",
                  "title": "Response List Resources Api V1 Resources Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/files": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Files",
        "description": "List files belonging to the authenticated user.",
        "operationId": "list_files_api_v1_files_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FileRef"
                  },
                  "type": "array",
                  "title": "Response List Files Api V1 Files Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Upload File",
        "description": "Create a text file from the JSON upload contract.",
        "operationId": "upload_file_api_v1_files_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileRef"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/files/content": {
      "put": {
        "tags": [
          "catalog"
        ],
        "summary": "Upload Streamed File",
        "description": "Stream and validate a text or PDF upload against storage limits.",
        "operationId": "upload_streamed_file_api_v1_files_content_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileRef"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/{file_id}/download": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Download File",
        "description": "Download the user's file with safe response headers.",
        "operationId": "download_file_api_v1_files__file_id__download_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/{file_id}": {
      "delete": {
        "tags": [
          "catalog"
        ],
        "summary": "Delete File",
        "description": "Delete one file belonging to the authenticated user.",
        "operationId": "delete_file_api_v1_files__file_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Artifacts",
        "description": "List AF3 artifact metadata visible to the user.",
        "operationId": "list_artifacts_api_v1_artifacts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactRef"
                  },
                  "type": "array",
                  "title": "Response List Artifacts Api V1 Artifacts Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}/preview": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Preview Artifact",
        "description": "Decode a small text artifact for an in-app preview.",
        "operationId": "preview_artifact_api_v1_artifacts__artifact_id__preview_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{artifact_id}/download": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Download Artifact",
        "description": "Download an owned AF3 artifact with safe response headers.",
        "operationId": "download_artifact_api_v1_artifacts__artifact_id__download_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactPreview": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "text"
        ],
        "title": "ArtifactPreview"
      },
      "ArtifactRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "available": {
            "type": "boolean",
            "title": "Available",
            "default": false
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind"
        ],
        "title": "ArtifactRef"
      },
...
      "CatalogItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CatalogItem"
      },
...
      "FileRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "status": {
            "type": "string",
            "const": "ready",
            "title": "Status",
            "default": "ready"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "size"
        ],
        "title": "FileRef"
      },
      "FileUploadRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "size": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Size"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "name",
          "size",
          "content"
        ],
        "title": "FileUploadRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}